+++ src/icons/iconComponent.jsx
@@ -46,10 +46,7 @@ export const Fb = () => (
 export const Telegram = () => (
   <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="icon">
     <rect x="0.5" y="0.5" width="23" height="23" rx="5.5" />
-    <path
-      d="M12.0949 13.0225L14.7923 1.29584C14.8113 1.2134 14.8074 1.12735 14.7811 1.04695C14.7547 0.966561 14.707 0.894878 14.6429 0.839628C14.5788 0.784291 14.5009 0.747499 14.4174 0.733222L12.0949 13.0225ZM12.0949 13.0225C12.0595 13.1795 11.9837 13.3244 11.8749 13.4429M12.0949 13.0225L11.8749 13.4429M11.8749 13.4429C11.766 13.5614 11.628 13.6493 11.4747 13.6979L11.8749 13.4429ZM11.526 2.33001L13.7147 1.47347L13.5169 1.61525L11.8043 2.84259L11.526 2.33001Z"
-      fill="#fff"
-    />
+    <path d="M18.8054 5.63676C18.7279 5.56987 18.6337 5.52538 18.5328 5.50808C18.4319 5.49079 18.3282 5.50136 18.2329 5.53864L5.56665 10.4955C5.38714 10.5653 5.23517 10.6915 5.1336 10.8552C5.03204 11.0188 4.98639 11.2111 5.00352 11.4029C5.02066 11.5947 5.09964 11.7758 5.2286 11.9189C5.35755 12.0619 5.52949 12.1592 5.71853 12.1961L9.0004 12.8405V16.4999C8.99976 16.6992 9.05899 16.8941 9.17043 17.0594C9.28187 17.2247 9.44037 17.3526 9.6254 17.4268C9.81015 17.5022 10.0133 17.5203 10.2085 17.4787C10.4037 17.4372 10.5818 17.3378 10.7198 17.1936L12.3023 15.5524L14.8129 17.7499C14.994 17.9106 15.2277 17.9995 15.4698 17.9999C15.5759 17.9998 15.6813 17.9831 15.7823 17.9505C15.9472 17.8982 16.0956 17.8035 16.2126 17.676C16.3296 17.5485 16.4112 17.3925 16.4492 17.2236L18.986 6.18739C19.0087 6.08785 19.0039 5.98402 18.9721 5.88702C18.9403 5.79002 18.8826 5.70351 18.8054 5.63676ZM14.9842 7.88676L9.38478 11.8968L6.28478 11.2886L14.9842 7.88676ZM10.0004 16.4999V13.5324L11.5498 14.8911L10.0004 16.4999ZM15.471 16.9999L10.3035 12.4686L17.741 7.13801L15.471 16.9999Z" />
   </svg>
 );
